Create SoftLink_Ex1
The issue:
The /home filesystem is nearly full and hence there is hardly any space left in it to store user files.
You being the system administrator have determined that there’s a lot of free space available in the "/opt" filesystem. Create a directory named "/opt/userdata" to store the users data.
To make access simple for all the users, create a soft link "/home/sdata" to "/opt/userdata".