Create Permission_Ex1
Tom is the Project Manager at XYZ.INC and is in charge of a group of developers who are about to develop an E-Commerce web application. Tom would like the developers to make use of a common directory in order to save their project work. He also wants to ensure that no one but the developers have access to the directory. Tom has obtained all the necessary approvals ahead of approaching you, the System Admin.
Your task: Create a common directory named "/opt/common" on the server as per the
following requirements.
 - The directory should be owned by user "prjmgr" and group "developers".
 - Make sure the "/opt/common" directory should have only full access to owner and group. Set the SETGID bit on, so that the same permission will be replicated from parent directory to all files & Sub-directories.
 - For verification, login or su as "prjmgr" user and create a file named "/opt/common/sgidfile". The file group should be "developers" group.