Create Package_Installation_Using_RPM_Command_EX2
Tom has emailed a software package to you which is in .zip format. In order to install the package, first of all you need the “unzip” command line utility. At the same time, you have found that the tcpdump utility that exists on the server is serving absolutely no purpose at all. Therefore your task is to...

1. Install the following package using the rpm command, the package is available in the /opt/Software Directory.
unzip-6.0-20.el7.x86_64.rpm
2. Uninstall the following package using the rpm command
Tcpdump