Create Creating_Repository_Updating_and_Installing_Packages_EX1
A while ago, one of our server users had raised a request for a software package. While trying to install the package you received the error that the package is not found in any of the currently configured yum repositories. You did a quick search online and found that this particular package is available in the Centos repository. Therefore your task is to ...

1. Create a software repository file "/etc/yum.repos.d/centos.repo"

        Configure the repository with the following parameters

        1. Repo-id = "centos"

        2. Repo-baseurl = "http://mirror.centos.org/centos-7/7/os/x86_64/"

        3. enable GPG signatures verification.

        Hint: Sample repository file example.repo is available in "/opt/Software/" directory.

 

2. Using the centos repository perform the following activities.

       1. Update centos-release package to the latest version

       2. Install centos-bookmarks and centos-indexhtml packages.